# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Financial Data API",
    "description": "주식 시장 데이터를 위한 종합 API - 뉴스, 공시, 어닝콜, AI 분석 리포트 제공",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/news": {
      "get": {
        "tags": ["News"],
        "summary": "모든 뉴스 목록 조회",
        "description": "시스템에 등록된 모든 뉴스 목록을 반환합니다.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "반환할 뉴스 개수 제한",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "건너뛸 뉴스 개수",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/News"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/news/{id}": {
      "get": {
        "tags": ["News"],
        "summary": "특정 뉴스 상세 조회",
        "description": "ID로 특정 뉴스의 상세 정보를 조회합니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "뉴스 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/News"
                }
              }
            }
          },
          "404": {
            "description": "뉴스를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/news/stock/{ticker}": {
      "get": {
        "tags": ["News"],
        "summary": "특정 종목 뉴스 조회",
        "description": "특정 종목의 모든 뉴스를 조회합니다.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "description": "종목 티커 심볼",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/News"
                      }
                    },
                    "ticker": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/filings": {
      "get": {
        "tags": ["Filings"],
        "summary": "모든 공시 목록 조회",
        "description": "시스템에 등록된 모든 공시 목록을 반환합니다.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "반환할 공시 개수 제한",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "건너뛸 공시 개수",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Filing"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/filings/{id}": {
      "get": {
        "tags": ["Filings"],
        "summary": "특정 공시 상세 조회",
        "description": "ID로 특정 공시의 상세 정보를 조회합니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "공시 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Filing"
                }
              }
            }
          },
          "404": {
            "description": "공시를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/filings/stock/{ticker}": {
      "get": {
        "tags": ["Filings"],
        "summary": "특정 종목 공시 조회",
        "description": "특정 종목의 모든 공시를 조회합니다.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "description": "종목 티커 심볼",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Filing"
                      }
                    },
                    "ticker": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/earnings-calls": {
      "get": {
        "tags": ["Earnings Calls"],
        "summary": "모든 어닝콜 목록 조회",
        "description": "시스템에 등록된 모든 어닝콜 목록을 반환합니다.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "반환할 어닝콜 개수 제한",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "건너뛸 어닝콜 개수",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EarningsCall"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/earnings-calls/{id}": {
      "get": {
        "tags": ["Earnings Calls"],
        "summary": "특정 어닝콜 상세 조회",
        "description": "ID로 특정 어닝콜의 상세 정보를 조회합니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "어닝콜 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningsCall"
                }
              }
            }
          },
          "404": {
            "description": "어닝콜을 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/earnings-calls/stock/{ticker}": {
      "get": {
        "tags": ["Earnings Calls"],
        "summary": "특정 종목 어닝콜 조회",
        "description": "특정 종목의 모든 어닝콜을 조회합니다.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "description": "종목 티커 심볼",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EarningsCall"
                      }
                    },
                    "ticker": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/reports": {
      "get": {
        "tags": ["AI Reports"],
        "summary": "모든 AI 리포트 목록 조회",
        "description": "시스템에 등록된 모든 AI 분석 리포트 목록을 반환합니다.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "반환할 리포트 개수 제한",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "건너뛸 리포트 개수",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AIReport"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/reports/{id}": {
      "get": {
        "tags": ["AI Reports"],
        "summary": "특정 AI 리포트 상세 조회",
        "description": "ID로 특정 AI 분석 리포트의 상세 정보를 조회합니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "리포트 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIReport"
                }
              }
            }
          },
          "404": {
            "description": "리포트를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/reports/stock/{ticker}": {
      "get": {
        "tags": ["AI Reports"],
        "summary": "특정 종목 AI 리포트 조회",
        "description": "특정 종목의 모든 AI 분석 리포트를 조회합니다.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "description": "종목 티커 심볼",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AIReport"
                      }
                    },
                    "ticker": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/stocks/{ticker}": {
      "get": {
        "tags": ["Stocks"],
        "summary": "특정 종목 정보 조회",
        "description": "티커 심볼로 특정 종목의 정보를 조회합니다.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "description": "종목 티커 심볼",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적인 응답",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stock"
                }
              }
            }
          },
          "404": {
            "description": "종목을 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "News": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "뉴스 고유 ID"
          },
          "stockInfo": {
            "$ref": "#/components/schemas/StockInfo"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "description": "뉴스 발행 시간"
          },
          "collectedAt": {
            "type": "string",
            "format": "date-time",
            "description": "뉴스 수집 시간"
          },
          "source": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "뉴스 소스명"
              },
              "url": {
                "type": "string",
                "format": "uri",
                "description": "뉴스 원문 URL"
              }
            }
          },
          "title": {
            "type": "object",
            "properties": {
              "original": {
                "type": "string",
                "description": "원문 제목"
              },
              "translated": {
                "type": "string",
                "description": "번역된 제목"
              }
            }
          },
          "summary": {
            "type": "object",
            "properties": {
              "oneLiner": {
                "type": "string",
                "description": "한 줄 요약"
              },
              "content": {
                "type": "string",
                "description": "상세 요약 내용"
              }
            }
          },
          "summary_translate": {
            "type": "object",
            "properties": {
              "oneLiner_ko": {
                "type": "string",
                "description": "한국어 한 줄 요약"
              },
              "content_ko": {
                "type": "string",
                "description": "한국어 상세 요약"
              }
            }
          },
          "analysis": {
            "type": "object",
            "properties": {
              "sentiment": {
                "type": "object",
                "properties": {
                  "score": {
                    "type": "number",
                    "description": "감정 점수 (-1 ~ 1)"
                  },
                  "label": {
                    "type": "string",
                    "description": "감정 라벨"
                  }
                }
              },
              "keywords": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "키워드 목록"
              },
              "relatedStocks": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "관련 종목 티커 목록"
              }
            }
          }
        }
      },
      "Filing": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "공시 고유 ID"
          },
          "stockInfo": {
            "$ref": "#/components/schemas/StockInfo"
          },
          "filingDate": {
            "type": "string",
            "format": "date",
            "description": "공시 제출일"
          },
          "periodOfReport": {
            "type": "string",
            "format": "date",
            "description": "보고 대상 기간 종료일"
          },
          "filingType": {
            "type": "string",
            "description": "공시 종류 (10-K, 10-Q, 8-K 등)"
          },
          "title": {
            "type": "object",
            "properties": {
              "original": {
                "type": "string",
                "description": "원문 제목"
              },
              "translated": {
                "type": "string",
                "description": "번역된 제목"
              }
            }
          },
          "documentUrl": {
            "type": "string",
            "format": "uri",
            "description": "원본 문서 링크"
          },
          "content": {
            "type": "object",
            "properties": {
              "riskFactors": {
                "type": "object",
                "properties": {
                  "original_full": {
                    "type": "string",
                    "description": "원문 전체 내용"
                  },
                  "summary_original": {
                    "type": "string",
                    "description": "원문 요약"
                  },
                  "summary_translated": {
                    "type": "string",
                    "description": "번역된 요약"
                  }
                }
              },
              "managementDiscussion": {
                "type": "object",
                "properties": {
                  "original_full": {
                    "type": "string",
                    "description": "원문 전체 내용"
                  },
                  "summary_original": {
                    "type": "string",
                    "description": "원문 요약"
                  },
                  "summary_translated": {
                    "type": "string",
                    "description": "번역된 요약"
                  }
                }
              }
            }
          }
        }
      },
      "EarningsCall": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "어닝콜 고유 ID"
          },
          "stockInfo": {
            "$ref": "#/components/schemas/StockInfo"
          },
          "callTimestamp": {
            "type": "string",
            "format": "date-time",
            "description": "어닝콜 시작 시각"
          },
          "fiscalPeriod": {
            "type": "string",
            "description": "회계 기간"
          },
          "highlights": {
            "type": "object",
            "properties": {
              "overallSummary": {
                "type": "object",
                "properties": {
                  "original": {
                    "type": "string",
                    "description": "원문 전체 요약"
                  },
                  "translated": {
                    "type": "string",
                    "description": "번역된 전체 요약"
                  }
                }
              },
              "keyQnA": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "question": {
                      "type": "string",
                      "description": "질문"
                    },
                    "answer": {
                      "type": "string",
                      "description": "답변"
                    },
                    "questioner": {
                      "type": "string",
                      "description": "질문자"
                    }
                  }
                }
              },
              "toneAnalysis": {
                "type": "object",
                "properties": {
                  "overall": {
                    "type": "string",
                    "description": "전체적인 톤"
                  },
                  "ceoTone": {
                    "type": "string",
                    "description": "CEO 톤"
                  },
                  "cfoTone": {
                    "type": "string",
                    "description": "CFO 톤"
                  }
                }
              }
            }
          },
          "transcript": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "sequence": {
                  "type": "integer",
                  "description": "순서"
                },
                "speaker": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "발언자 이름"
                    },
                    "role": {
                      "type": "string",
                      "description": "발언자 역할"
                    }
                  }
                },
                "text": {
                  "type": "object",
                  "properties": {
                    "original": {
                      "type": "string",
                      "description": "원문"
                    },
                    "translated": {
                      "type": "string",
                      "description": "번역문"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "Stock": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string",
            "description": "종목 티커 심볼"
          },
          "name": {
            "type": "string",
            "description": "회사명"
          },
          "exchange": {
            "type": "string",
            "description": "거래소"
          },
          "industry": {
            "type": "string",
            "description": "업종"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "에러 메시지"
          },
          "message": {
            "type": "string",
            "description": "상세 에러 설명"
          }
        }
      }
    }
  }
}